- clean up integer types, no warnings about signedness anymore - replace 127.0.0.1 with localhost to hopefully work around debian test issue - make SHA implementations endian aware - add date_time library as a dependency

+++ lib-xml/src/xpath.cpp
@@ -2197,7 +2197,7 @@ Token xpath_imp::get_next_token()
 					}
 					else
 					{
-						for (int i = 0; i < kCoreFunctionCount; ++i)
+						for (size_t i = 0; i < kCoreFunctionCount; ++i)
 						{
 							if (m_token_string == kCoreFunctionInfo[i].name)
 							{
